meti todas las funciones a functions.js, hago el sprint antes de empezar a mudar todo a la api
