+++ JS/past.js
@@ -1,41 +1,49 @@
 const eventos = data.events;
 
+import { crearCheckbox, filtrarEventosCategoria, filtrarEventosSearch, aplicarFiltros, mostrarMensaje, crearCardPast} from '../Modules/functions.js';
+
 // Cards
 let currentDate = data.currentDate;
 currentDate = currentDate.split("-");
 
 const cardsContainer = document.getElementById("cardsContainer")
 
-function compararFecha(fecha) {
-    let fechaSplit = fecha.split("-");
-    for (let i = 0; i < 3; i++) {
-        if (parseInt(currentDate[i]) < parseInt(fechaSplit[i])) {
-            return "mayor";
-        } else if (parseInt(currentDate[i]) > parseInt(fechaSplit[i])) {
-            return "menor";
-        }
-    }
-    return "igual";
+crearCardPast(eventos, cardsContainer, currentDate)
+
+// Checkbox
+
+let checkboxContainer = document.querySelector("#checkbox");
+
+crearCheckbox(eventos, checkboxContainer)
+
+// filtro por checkbox
+
+let checkboxEvent = document.getElementById("checkbox")
+//let checkboxCategory = document.querySelector(".checkboxLabel")
+let categoryFilter = ".category_filter"
+
+checkboxEvent.addEventListener("change", function(){
+  crearCardPast(filtrarEventosCategoria(eventos, categoryFilter), cardsContainer, currentDate)
+})
+
+// filtro por search bar y cruzado
+
+let searchBar = document.getElementById("searchBar")
+let searchButton = document.getElementById("searchButton")
+
+
+let nameEventos = []
+for (const evento of eventos) {
+  nameEventos.push(evento.name)
 }
 
-function crearCard(eventos) {
-    let card = "";
-    for (const evento of eventos) {
-        if(compararFecha(evento.date) == "menor"){
-            card += `<div class="card" style="width: 18rem;">
-            <div class="card-body d-flex flex-column justify-content-center align-items-center">
-                <img src="${evento.image}" alt="Food fair" class="card-img">
-                <h5 class="card-title">${evento.name}</h5>
-                <p class="card-text">${evento.description}</p>
-                <div class="d-flex justify-content-around align-items-baseline gap-4">
-                    <p>$${evento.price}</p>
-                    <button type="button" class="btn btn-primary"><a href="Details.html?id=${evento._id}" class="card-link color-details">Details</a></button>                  
-                </div>
-                </div>
-            </div>`
-        }
-    }
-    cardsContainer.innerHTML = card;
-}
-
-crearCard(eventos)
+searchButton.addEventListener("click", function (event) {
+  event.preventDefault()
+  let eventosFiltrados = aplicarFiltros(eventos, categoryFilter, nameEventos, searchBar);
+  if(eventosFiltrados.length === 0){
+    mostrarMensaje();
+  }
+  else{
+    crearCardPast(eventosFiltrados, cardsContainer, currentDate);
+  }
+})
