+++ JS/upcoming.js
@@ -1,5 +1,7 @@
 const eventos = data.events;
 
+import { crearCardUpcoming, crearCheckbox, filtrarEventosCategoria, filtrarEventosSearch, aplicarFiltros, mostrarMensaje } from '../Modules/functions.js';
+
 // Cards
 
 var currentDate = data.currentDate;
@@ -7,39 +9,39 @@ currentDate = currentDate.split("-");
 
 const cardsContainer = document.getElementById("cardsContainer")
 
+crearCardUpcoming(eventos, cardsContainer, currentDate)
 
+// Checkbox
 
-function compararFecha(fecha) {
-    let fechaSplit = fecha.split("-");
-    for (let i = 0; i < 3; i++) {
-        if (parseInt(currentDate[i]) < parseInt(fechaSplit[i])) {
-            return "mayor";
-        } else if (parseInt(currentDate[i]) > parseInt(fechaSplit[i])) {
-            return "menor";
-        }
-    }
-    return "igual";
+let checkboxContainer = document.querySelector("#checkbox");
+
+crearCheckbox(eventos, checkboxContainer)
+
+let checkboxEvent = document.getElementById("checkbox")
+let checkboxCategory = document.querySelector(".checkboxLabel")
+let categoryFilter = ".category_filter"
+
+checkboxEvent.addEventListener("change", function(){
+  crearCardUpcoming(filtrarEventosCategoria(eventos, categoryFilter), cardsContainer, currentDate)
+})
+
+let searchBar = document.getElementById("searchBar")
+let searchButton = document.getElementById("searchButton")
+
+//funcion que reciba eventos y filtre por nombre
+
+let nameEventos = []
+for (const evento of eventos) {
+  nameEventos.push(evento.name)
 }
 
-function crearCard(eventos){
-    let card = "";
-    for (const evento of eventos) {
-        if(compararFecha(evento.date) === "mayor"){
-            card += `<div class="card" style="width: 18rem;">
-                        <div class="card-body d-flex flex-column justify-content-center align-items-center">
-                            <img src="${evento.image}" alt="Food fair" class="card-img">
-                            <h5 class="card-title">${evento.name}</h5>
-                            <p class="card-text">${evento.description}</p>
-                            <div class="d-flex justify-content-around align-items-baseline gap-4">
-                                <p>$${evento.price}</p>
-                                <button type="button" class="btn btn-primary"><a href="Details.html?id=${evento._id}" class="card-link color-details">Details</a></button>                  
-                            </div>
-                        </div>
-                    </div>`
-        }
-    }
-
-    cardsContainer.innerHTML = card;
-}
-
-crearCard(eventos)
+searchButton.addEventListener("click", function (event) {
+  event.preventDefault()
+  let eventosFiltrados = aplicarFiltros(eventos, categoryFilter, nameEventos, searchBar);
+  if(eventosFiltrados.length === 0){
+    mostrarMensaje();
+  }
+  else{
+    crearCardUpcoming(eventosFiltrados, cardsContainer, currentDate);
+  }
+})
