+++ WebPages/Index.html
@@ -101,6 +101,6 @@
     </footer>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm" crossorigin="anonymous"></script>
     <script src="../JS/data.js"></script>
-    <script src="../JS/main.js"></script>
+    <script src="../JS/main.js" type="module"></script>
 </body>
 </html>
