+++ WebPages/UpcomingEvents.html
@@ -74,45 +74,16 @@
 
           <div>
             <nav class="navbar bg-body-tertiary mt-5 justify-content-center">
-                <div class="container-fluid col-lg-8 d-flex justify-content-start gap-4">
-                  <div class="form-check">
-                    <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault">
-                    <label class="form-check-label" for="flexCheckDefault">
-                      <p>Food Fair</p>
-                    </label>
-                  </div>
-                  <div class="form-check">
-                    <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault">
-                    <label class="form-check-label" for="flexCheckDefault">
-                      <p>Museum</p>
-                    </label>
-                  </div>
-                  <div class="form-check">
-                    <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault">
-                    <label class="form-check-label" for="flexCheckDefault">
-                      <p>Costume Party</p>
-                    </label>
-                  </div>
-                  <div class="form-check">
-                    <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault">
-                    <label class="form-check-label" for="flexCheckDefault">
-                      <p>Music Concert</p>
-                    </label>
-                  </div>
-                  <div class="form-check">
-                    <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault">
-                    <label class="form-check-label" for="flexCheckDefault">
-                      <p>Race</p>
-                    </label>
-                  </div>
-                </div>
-                <div class="col-lg-3 d-flex justify-content-center">
-                  <form class="d-flex" role="search">
-                    <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search">
-                    <button class="btn btn-outline-success" type="submit">Search</button>
-                  </form>
-                </div>
-              </nav>
+              <div class="container-fluid col-lg-8 d-flex justify-content-start gap-4" id="checkbox">
+
+              </div>
+              <div class="col-lg-3 d-flex justify-content-center">
+                <form class="d-flex" role="search">
+                  <input class="form-control me-2" id="searchBar" type="search" placeholder="Search" aria-label="Search">
+                  <button class="btn btn-outline-success" id="searchButton" type="submit">Search</button>
+                </form>
+              </div>
+            </nav>
           </div>
 
           <div class="d-flex flex-direction-row justify-content-around mt-5 bg-body-tertiary flex-wrap gap-3" id="cardsContainer">
@@ -131,6 +102,6 @@
     </footer>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm" crossorigin="anonymous"></script>
     <script src="../JS/data.js"></script>
-    <script src="../JS/upcoming.js"></script>
+    <script src="../JS/upcoming.js" type="module"></script>
 </body>
 </html>
